Fix incompatibility with FML 3.x+ so that Forge-based MC1.3+ mod packs are possible
